push

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ahmaidi <ahmaidi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/01 20:45:25 by ahmaidi           #+#    #+#             */
-/*   Updated: 2021/12/02 22:36:08 by ahmaidi          ###   ########.fr       */
+/*   Updated: 2021/12/05 00:48:30 by ahmaidi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	ft_printf(const char *format,	...);
 int	ft_putchar(char c);
 int	ft_putstr(char *s);
 int	ft_putnbr(int n);
-int	ft_putunbr(long i);
+int	ft_putunbr(unsigned int i);
 int	ft_hexa(unsigned int i, char base);
 
 #endif
